refactor: add pyproject.toml
