+++ .github/workflows/lint.yml
@@ -5,17 +5,17 @@ jobs:
   lint:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
-    - uses: actions/setup-python@v5
-      with:
-        python-version: '3.10'
-    - name: Install requirements_dev.txt
-      run: pip install -r requirements_dev.txt
-    - name: Check isort
-      run: isort --check-only iatidata/
-    - name: Check black
-      run: black --check iatidata/
-    - name: Check flake8
-      run: flake8 iatidata/
-    - name: Check types with mypy
-      run: mypy iatidata/
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install requirements_dev.txt
+        run: pip install -r requirements_dev.txt
+      - name: Check isort
+        run: isort --check-only iati_tables/
+      - name: Check black
+        run: black --check iati_tables/
+      - name: Check flake8
+        run: flake8 iati_tables/
+      - name: Check types with mypy
+        run: mypy iati_tables/
