+++ .github/workflows/test.yml
@@ -5,11 +5,13 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
-    - uses: actions/setup-python@v5
-      with:
-        python-version: '3.10'
-    - name: Install requirements_dev.txt
-      run: pip install -r requirements_dev.txt
-    - name: Test
-      run: python -m pytest iatidata/
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+      - name: Install requirements_dev.txt
+        run: pip install -r requirements_dev.txt
+      - name: Install local package
+        run: pip install .
+      - name: Test
+        run: pytest
