+++ pyproject.toml
@@ -1,0 +1,36 @@
+[build-system]
+requires = ["flit_core >=3.2,<4"]
+build-backend = "flit_core.buildapi"
+
+[project]
+name = "iati-tables"
+description = "IATI Tables"
+version = "1.0.0"
+readme = "README.md"
+dependencies = [
+  "boto3",
+  "click",
+  "fastavro",
+  "google-cloud-bigquery",
+  "iatikit",
+  "lxml",
+  "psycopg2-binary",
+  "sqlalchemy",
+  "xmlschema",
+]
+
+[project.optional-dependencies]
+dev = [
+  "black",
+  "flake8",
+  "isort",
+  "mypy",
+  "pytest",
+  "sphinx",
+  "sphinx-togglebutton",
+  "sphinx-intl",
+  "sphinx_rtd_theme",
+  "sphinx-autobuild",
+  "types-requests",
+  "types-lxml",
+]
