--- requirements.in
@@ -1,9 +1,0 @@
-lxml
-xmlschema
-sqlalchemy
-psycopg2-binary
-click
-boto3
-iatikit
-fastavro
-google-cloud-bigquery
