+++ requirements.txt
@@ -2,10 +2,10 @@
 # This file is autogenerated by pip-compile with Python 3.10
 # by the following command:
 #
-#    pip-compile requirements.in
+#    pip-compile --output-file=requirements.txt --strip-extras pyproject.toml
 #
 boto3==1.34.19
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 botocore==1.34.19
     # via
     #   boto3
@@ -17,13 +17,13 @@ certifi==2023.11.17
 charset-normalizer==3.3.2
     # via requests
 click==8.1.7
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 configparser==6.0.0
     # via iatikit
 elementpath==4.1.5
     # via xmlschema
 fastavro==1.9.3
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 google-api-core==2.15.0
     # via
     #   google-cloud-bigquery
@@ -33,7 +33,7 @@ google-auth==2.26.2
     #   google-api-core
     #   google-cloud-core
 google-cloud-bigquery==3.16.0
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 google-cloud-core==2.4.1
     # via google-cloud-bigquery
 google-crc32c==1.5.0
@@ -43,7 +43,7 @@ google-resumable-media==2.7.0
 googleapis-common-protos==1.62.0
     # via google-api-core
 iatikit==3.4.0
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 idna==3.6
     # via requests
 jmespath==1.0.1
@@ -52,7 +52,7 @@ jmespath==1.0.1
     #   botocore
 lxml==5.1.0
     # via
-    #   -r requirements.in
+    #   iati-tables (pyproject.toml)
     #   iatikit
 packaging==23.2
     # via google-cloud-bigquery
@@ -61,7 +61,7 @@ protobuf==4.25.2
     #   google-api-core
     #   googleapis-common-protos
 psycopg2-binary==2.9.9
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 pyasn1==0.5.1
     # via
     #   pyasn1-modules
@@ -84,7 +84,7 @@ s3transfer==0.10.0
 six==1.16.0
     # via python-dateutil
 sqlalchemy==2.0.31
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
 typing-extensions==4.12.2
     # via sqlalchemy
 urllib3==2.0.7
@@ -92,4 +92,4 @@ urllib3==2.0.7
     #   botocore
     #   requests
 xmlschema==3.0.1
-    # via -r requirements.in
+    # via iati-tables (pyproject.toml)
