--- requirements_dev.in
@@ -1,14 +1,0 @@
--r requirements.txt
-
-pytest
-black
-isort
-flake8
-mypy
-types-requests
-types-lxml
-sphinx
-sphinx-togglebutton
-sphinx-intl
-sphinx_rtd_theme
-sphinx-autobuild
