+++ requirements_dev.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.10
 # by the following command:
 #
-#    pip-compile requirements_dev.in
+#    pip-compile --extra=dev --output-file=requirements_dev.txt --strip-extras pyproject.toml
 #
 alabaster==0.7.16
     # via sphinx
@@ -11,37 +11,28 @@ babel==2.14.0
     #   sphinx
     #   sphinx-intl
 black==23.12.1
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 boto3==1.34.19
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 botocore==1.34.19
     # via
-    #   -r requirements.txt
     #   boto3
     #   s3transfer
 cachetools==5.3.2
-    # via
-    #   -r requirements.txt
-    #   google-auth
+    # via google-auth
 certifi==2023.11.17
-    # via
-    #   -r requirements.txt
-    #   requests
+    # via requests
 charset-normalizer==3.3.2
-    # via
-    #   -r requirements.txt
-    #   requests
+    # via requests
 click==8.1.7
     # via
-    #   -r requirements.txt
     #   black
+    #   iati-tables (pyproject.toml)
     #   sphinx-intl
 colorama==0.4.6
     # via sphinx-autobuild
 configparser==6.0.0
-    # via
-    #   -r requirements.txt
-    #   iatikit
+    # via iatikit
 cssselect==1.2.0
     # via types-lxml
 docutils==0.20.1
@@ -50,81 +41,65 @@ docutils==0.20.1
     #   sphinx-rtd-theme
     #   sphinx-togglebutton
 elementpath==4.1.5
-    # via
-    #   -r requirements.txt
-    #   xmlschema
+    # via xmlschema
 exceptiongroup==1.2.0
     # via pytest
 fastavro==1.9.3
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 flake8==7.0.0
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 google-api-core==2.15.0
     # via
-    #   -r requirements.txt
     #   google-cloud-bigquery
     #   google-cloud-core
 google-auth==2.26.2
     # via
-    #   -r requirements.txt
     #   google-api-core
     #   google-cloud-core
 google-cloud-bigquery==3.16.0
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 google-cloud-core==2.4.1
-    # via
-    #   -r requirements.txt
-    #   google-cloud-bigquery
+    # via google-cloud-bigquery
 google-crc32c==1.5.0
-    # via
-    #   -r requirements.txt
-    #   google-resumable-media
+    # via google-resumable-media
 google-resumable-media==2.7.0
-    # via
-    #   -r requirements.txt
-    #   google-cloud-bigquery
+    # via google-cloud-bigquery
 googleapis-common-protos==1.62.0
-    # via
-    #   -r requirements.txt
-    #   google-api-core
+    # via google-api-core
 iatikit==3.4.0
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 idna==3.6
-    # via
-    #   -r requirements.txt
-    #   requests
+    # via requests
 imagesize==1.4.1
     # via sphinx
 iniconfig==2.0.0
     # via pytest
 isort==5.13.2
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 jinja2==3.1.3
     # via sphinx
 jmespath==1.0.1
     # via
-    #   -r requirements.txt
     #   boto3
     #   botocore
 livereload==2.6.3
     # via sphinx-autobuild
 lxml==5.1.0
     # via
-    #   -r requirements.txt
+    #   iati-tables (pyproject.toml)
     #   iatikit
 markupsafe==2.1.5
     # via jinja2
 mccabe==0.7.0
     # via flake8
 mypy==1.8.0
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 mypy-extensions==1.0.0
     # via
     #   black
     #   mypy
 packaging==23.2
     # via
-    #   -r requirements.txt
     #   black
     #   google-cloud-bigquery
     #   pytest
@@ -137,20 +112,16 @@ pluggy==1.4.0
     # via pytest
 protobuf==4.25.2
     # via
-    #   -r requirements.txt
     #   google-api-core
     #   googleapis-common-protos
 psycopg2-binary==2.9.9
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 pyasn1==0.5.1
     # via
-    #   -r requirements.txt
     #   pyasn1-modules
     #   rsa
 pyasn1-modules==0.3.0
-    # via
-    #   -r requirements.txt
-    #   google-auth
+    # via google-auth
 pycodestyle==2.11.1
     # via flake8
 pyflakes==3.2.0
@@ -158,50 +129,43 @@ pyflakes==3.2.0
 pygments==2.17.2
     # via sphinx
 pytest==7.4.4
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 python-dateutil==2.8.2
     # via
-    #   -r requirements.txt
     #   botocore
     #   google-cloud-bigquery
 requests==2.31.0
     # via
-    #   -r requirements.txt
     #   google-api-core
     #   google-cloud-bigquery
     #   iatikit
     #   sphinx
 rsa==4.9
-    # via
-    #   -r requirements.txt
-    #   google-auth
+    # via google-auth
 s3transfer==0.10.0
-    # via
-    #   -r requirements.txt
-    #   boto3
+    # via boto3
 six==1.16.0
     # via
-    #   -r requirements.txt
     #   livereload
     #   python-dateutil
 snowballstemmer==2.2.0
     # via sphinx
 sphinx==7.2.6
     # via
-    #   -r requirements_dev.in
+    #   iati-tables (pyproject.toml)
     #   sphinx-autobuild
     #   sphinx-intl
     #   sphinx-rtd-theme
     #   sphinx-togglebutton
     #   sphinxcontrib-jquery
 sphinx-autobuild==2024.2.4
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 sphinx-intl==2.1.0
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 sphinx-rtd-theme==2.0.0
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 sphinx-togglebutton==0.3.2
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 sphinxcontrib-applehelp==1.0.8
     # via sphinx
 sphinxcontrib-devhelp==1.0.6
@@ -217,7 +181,7 @@ sphinxcontrib-qthelp==1.0.7
 sphinxcontrib-serializinghtml==1.1.10
     # via sphinx
 sqlalchemy==2.0.31
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 tomli==2.0.1
     # via
     #   black
@@ -230,26 +194,24 @@ types-beautifulsoup4==4.12.0.20240511
 types-html5lib==1.1.11.20240228
     # via types-beautifulsoup4
 types-lxml==2024.4.14
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 types-requests==2.31.0.20240125
-    # via -r requirements_dev.in
+    # via iati-tables (pyproject.toml)
 typing-extensions==4.12.2
     # via
-    #   -r requirements.txt
     #   black
     #   mypy
     #   sqlalchemy
     #   types-lxml
 urllib3==2.0.7
     # via
-    #   -r requirements.txt
     #   botocore
     #   requests
     #   types-requests
 wheel==0.42.0
     # via sphinx-togglebutton
 xmlschema==3.0.1
-    # via -r requirements.txt
+    # via iati-tables (pyproject.toml)
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
